requirements.txt
agora-token-builder==1.0.0
asgiref==3.7.1
attrs==23.1.0
autobahn==23.6.1
Automat==22.10.0
bcrypt==4.0.1
boto3==1.26.160
botocore==1.29.160
certifi==2023.5.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
constantly==15.1.0
crispy-bootstrap5==0.7
cryptography==41.0.1
daphne==4.0.0
distlib==0.3.6
Django==4.2.1
django-cors-headers==4.1.0
django-crispy-forms==2.0
django-storages==1.13.2
djangorestframework==3.14.0
filelock==3.12.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
jmespath==1.0.1
msgpack==1.0.5
paramiko==3.2.0
platformdirs==3.5.1
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyNaCl==1.5.0
pyOpenSSL==23.2.0
python-dateutil==2.8.2
pytz==2023.3
redis==4.5.5
s3transfer==0.6.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.4
storage==0.0.4.3
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.6.3
urllib3==1.26.16
virtualenv==20.23.0
zope.interface==6.0
